+++ src/functors.rs
@@ -7,7 +7,8 @@ use std::collections::{HashMap, HashSet};
 #[cfg(feature = "std")]
 use std::hash::BuildHasher;
 
-use num_traits::NumAssign;
+#[cfg(feature = "std")]
+use num::One;
 
 use crate::{Functor, Inner};
 
@@ -27,9 +28,7 @@ impl<I: Inner> Functor<I> for Vec<I> {
 }
 
 #[cfg(feature = "std")]
-impl<I: Inner, N: Clone + Default + NumAssign, S: BuildHasher + Default> Functor<I>
-    for HashMap<I, N, S>
-{
+impl<I: Inner, N: Clone + Default + One, S: BuildHasher + Default> Functor<I> for HashMap<I, N, S> {
     #[inline]
     fn pure(i: I) -> Self {
         let mut hm = Self::default();