+++ tests/fmap_flat.rs
@@ -32,7 +32,7 @@ fn test_flat_map_enumerator() {
     let mut rng = ChaCha8Rng::seed_from_u64(0);
     let output = random_process::<Enumerator>(&mut rng, 9);
 
-    assert_eq!(output.len(), 8 * 11 + 9);
+    assert_eq!(output.len(), 17 * 11);
 
     let counts = output.iter().fold(HashMap::new(), |mut map, d| {
         *map.entry(d).or_insert(0usize) += 1;
@@ -40,7 +40,7 @@ fn test_flat_map_enumerator() {
     });
 
     for i in 1..=9 {
-        assert_eq!(counts[&i], 1);
+        assert_eq!(counts[&i], 11);
     }
     for i in 200..=210 {
         assert_eq!(counts[&i], 8);
@@ -53,10 +53,10 @@ fn test_flat_map_counter() {
     let output = random_process::<Counter>(&mut rng, 9);
 
     assert_eq!(output.len(), 20);
-    assert_eq!(output.values().sum::<usize>(), 8 * 11 + 9);
+    assert_eq!(output.values().sum::<usize>(), 17 * 11);
 
     for i in 1..=9 {
-        assert_eq!(output[&i], 1);
+        assert_eq!(output[&i], 11);
     }
     for i in 200..=210 {
         assert_eq!(output[&i], 8);